Refactor: Fix changes per @pc494
