+++ CHANGELOG.rst
@@ -28,6 +28,13 @@ Deleted
 -------
 - Several expired methods of `Diffraction2D` associated with radial integration (#998)
 
+Changed
+-------
+- Subpixel refinement now a function of the :py:class:`~pyxem.signals.DiffractionVectors` class (#980)
+- The :py:class:`~pyxem.generators.SubpixelrefinementGenerator` class has been deprecated (#980)
+
+
+
 2024-01-05 - version 0.17.0
 ===========================
 Added
@@ -80,10 +87,7 @@ Fixed
   passing orientations as tuples.
 - Fix bug in calculating strain (#958)
 
-Changed
--------
-- Subpixel refinement now a function of the :py:class:`~pyxem.signals.DiffractionVectors` class (#980)
-- The :py:class:`~pyxem.generators.SubpixelrefinementGenerator` class has been deprecated (#980)
+
 
 
 2023-05-08 - version 0.15.1
