+++ pyxem/tests/signals/test_diffraction_vectors.py
@@ -25,20 +25,9 @@ from skimage.morphology import disk as disk2
 from hyperspy.signals import Signal2D
 from hyperspy.signal import BaseSignal
 
-from pyxem.signals import (
-    DiffractionVectors,
-    DiffractionVectors2D,
-    DiffractionVectors1D,
-    PolarVectors,
-)
-from pyxem.utils.subpixel_utils import (
-    _center_of_mass_hs,
-    get_experimental_square,
-    _conventional_xc,
-)
-import pyxem.dummy_data.make_diffraction_test_data as mtd
-from pyxem.signals import DiffractionVectors, DiffractionVectors2D, DiffractionVectors1D
-from pyxem.utils.subpixel_utils import (
+
+from pyxem.signals import DiffractionVectors, DiffractionVectors2D
+from pyxem.utils._subpixel_utils import (
     _center_of_mass_hs,
     get_experimental_square,
     _conventional_xc,
